improved routing and Header / Footer
fixed pagination bug
