+++ src/components/index.ts
@@ -1,2 +1,4 @@
+export { Home } from './Home';
+export { ErrorPage } from './ErrorPage';
 export { OwnerList } from './OwnerList';
 export { VetList } from './VetList';
