+++ src/navigation/Footer.tsx
@@ -1,0 +1,12 @@
+
+export default function Footer() {
+  return (
+    <div className="container-fluid">
+        <div className="row">
+          <div className="col-12 text-center">
+            <img src="/images/spring-pivotal-logo.png" alt="Sponsored by Pivotal" />
+          </div>
+        </div>
+      </div>
+  );
+}
