+++ src/navigation/Menu.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 
 const MenuItem = ({
   active,
