+++ src/navigation/index.ts
@@ -1,1 +1,2 @@
 export { default as Menu } from './Menu';
+export { default as Footer } from './Footer';
